project-04-mcp-enterprise-gateway/pyproject.toml
[project]
name = "mcp-enterprise-gateway"
version = "0.1.0"
description = "FastMCP composable gateway with Redis cache, rate limiting, and OpenTelemetry"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.0.0",
    "mcp>=1.5.0",
    "redis[hiredis]>=5.0.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.3.0",
    "boto3>=1.34.0",
    "httpx>=0.27.0",
    "openai>=1.50.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
